myvenv/requirements.txt
streamlit==1.8.0
openai==0.10.2
python-docx==0.8.11